feat: Improve UX for running tests
