+++ src/foundry.ts
@@ -4,16 +4,9 @@ import * as path from 'path';
 import { parse } from 'smol-toml';
 import * as vscode from 'vscode';
 import { getKaasBaseUrl } from './config';
-import {
-  getGitInfo,
-  getGitRepository,
-  gitApi,
-  hasUnpushedChanges,
-  hasWorkingTreeChanges,
-} from './git';
+import { GitInfo } from './git';
 import { components, JobKind, paths } from './kaas-api';
 import { pollForJobStatus } from './kaas_jobs';
-import { verifyVaultExists } from './kaas_vault';
 import { TestRunState } from './test_run_state';
 
 interface FoundryTest {
@@ -194,7 +187,8 @@ export async function runFoundryTestViaKaaS(
   testController: vscode.TestController,
   testRun: vscode.TestRun,
   test: vscode.TestItem,
-  testRunState: TestRunState
+  testRunState: TestRunState,
+  validatedGitInfo: GitInfo
 ): Promise<void> {
   console.log(`Processing Foundry test: ${test.id}`);
   test.busy = true;
@@ -207,65 +201,9 @@ export async function runFoundryTestViaKaaS(
     return;
   }
 
-  // --- Dirty git check ---
-  const git = gitApi();
-  const repository = await getGitRepository(git, worksaceFolder);
-  if (repository) {
-    const workingTreeChanges = await hasWorkingTreeChanges(repository);
-    const unpushedChanges = await hasUnpushedChanges(repository);
-    if (workingTreeChanges || unpushedChanges) {
-      const proceed = await vscode.window.showWarningMessage(
-        'You have uncommitted or unpushed changes. Please commit and push your changes before running a remote job on KaaS.',
-        { modal: true },
-        'Proceed Anyway',
-        'Cancel'
-      );
-      if (proceed !== 'Proceed Anyway') {
-        test.busy = false;
-        testRun.errored(test, new vscode.TestMessage('Job cancelled due to dirty git state.'));
-        return;
-      }
-    }
-  }
+  console.log(`Using validated git info for ${test.id}:`, validatedGitInfo);
 
-  const gitInfo = await getGitInfo(worksaceFolder);
-  if (!gitInfo) {
-    vscode.window
-      .showErrorMessage(
-        'KaaS requires access to your remote repository. Please install the Runtime Verification GitHub App and grant access.',
-        'Install App'
-      )
-      .then(selection => {
-        if (selection === 'Install App') {
-          vscode.env.openExternal(
-            vscode.Uri.parse('https://github.com/apps/runtime-verification-inc')
-          );
-        }
-      });
-    console.error(`Could not get git info for ${test.uri.fsPath}`);
-    test.busy = false;
-    testRun.errored(
-      test,
-      new vscode.TestMessage(
-        'Could not determine git origin. Make sure you are in a git repository with a remote named "origin" and KaaS has access.'
-      )
-    );
-    return;
-  }
-  console.log(`Git info for ${test.id}:`, gitInfo);
-
-  const { owner: organizationName, repo: vaultName } = gitInfo;
-
-  const verificationError = await verifyVaultExists(client, organizationName, vaultName);
-  if (verificationError) {
-    console.error(
-      `Vault verification failed for ${organizationName}/${vaultName}: ${verificationError}`
-    );
-    test.busy = false;
-    testRun.errored(test, new vscode.TestMessage(verificationError));
-    return;
-  }
-  console.log(`Vault verified for ${test.id}`);
+  const { owner: organizationName, repo: vaultName } = validatedGitInfo;
 
   // This is a placeholder for the Foundry-specific test arguments.
   // We are setting the `match-test` argument, similar to how it was done for local runs.
@@ -278,7 +216,7 @@ export async function runFoundryTestViaKaaS(
   ];
 
   const body: components['schemas']['CreateJobDto'] = {
-    branch: gitInfo.branch,
+    branch: validatedGitInfo.branch,
     kind: JobKind.foundry,
     // --- These fields are based on the Kontrol implementation and may need adjustment for Foundry ---
     kontrolVersion: 'latest',
