+++ src/kontrol.ts
@@ -2,16 +2,9 @@ import { Client } from 'openapi-fetch';
 import { parse } from 'smol-toml';
 import * as vscode from 'vscode';
 import { getKaasBaseUrl } from './config';
-import {
-  getGitInfo,
-  getGitRepository,
-  gitApi,
-  hasUnpushedChanges,
-  hasWorkingTreeChanges,
-} from './git';
+import { getGitInfo, gitApi, GitInfo } from './git';
 import { components, JobKind, JobStatus, paths } from './kaas-api';
 import { getJobStatusByJobId, pollForJobStatus } from './kaas_jobs';
-import { verifyVaultExists } from './kaas_vault';
 import { TestRunState } from './test_run_state';
 
 interface KontrolToml {
@@ -77,7 +70,8 @@ export async function runKontrolProfileViaKaaS(
   testController: vscode.TestController,
   testRun: vscode.TestRun,
   test: vscode.TestItem,
-  testRunState: TestRunState
+  testRunState: TestRunState,
+  validatedGitInfo: GitInfo
 ): Promise<void> {
   console.log(`Processing Kontrol test: ${test.id}`);
   test.busy = true;
@@ -90,65 +84,9 @@ export async function runKontrolProfileViaKaaS(
     return;
   }
 
-  // --- Dirty git check ---
-  const git = gitApi();
-  const repository = await getGitRepository(git, worksaceFolder);
-  if (repository) {
-    const workingTreeChanges = await hasWorkingTreeChanges(repository);
-    const unpushedChanges = await hasUnpushedChanges(repository);
-    if (workingTreeChanges || unpushedChanges) {
-      const proceed = await vscode.window.showWarningMessage(
-        'You have uncommitted or unpushed changes. Please commit and push your changes before running a remote job on KaaS.',
-        { modal: true },
-        'Proceed Anyway',
-        'Cancel'
-      );
-      if (proceed !== 'Proceed Anyway') {
-        test.busy = false;
-        testRun.errored(test, new vscode.TestMessage('Job cancelled due to dirty git state.'));
-        return;
-      }
-    }
-  }
-
-  const gitInfo = await getGitInfo(worksaceFolder);
-  if (!gitInfo) {
-    vscode.window
-      .showErrorMessage(
-        'KaaS requires access to your remote repository. Please install the Runtime Verification GitHub App and grant access.',
-        'Install App'
-      )
-      .then(selection => {
-        if (selection === 'Install App') {
-          vscode.env.openExternal(
-            vscode.Uri.parse('https://github.com/apps/runtime-verification-inc')
-          );
-        }
-      });
-    console.error(`Could not get git info for ${test.uri.fsPath}`);
-    test.busy = false;
-    testRun.errored(
-      test,
-      new vscode.TestMessage(
-        'Could not determine git origin. Make sure you are in a git repository with a remote named "origin" and KaaS has access.'
-      )
-    );
-    return;
-  }
-  console.log(`Git info for ${test.id}:`, gitInfo);
-
-  const { owner: organizationName, repo: vaultName } = gitInfo;
-
-  const verificationError = await verifyVaultExists(client, organizationName, vaultName);
-  if (verificationError) {
-    console.error(
-      `Vault verification failed for ${organizationName}/${vaultName}: ${verificationError}`
-    );
-    test.busy = false;
-    testRun.errored(test, new vscode.TestMessage(verificationError));
-    return;
-  }
-  console.log(`Vault verified for ${test.id}`);
+  // Git info and vault verification are now handled in the parent runTests function
+  const { owner: organizationName, repo: vaultName, branch } = validatedGitInfo;
+  console.log(`Using validated git info for ${test.id}:`, validatedGitInfo);
 
   let profiles: components['schemas']['CreateProveProfileDto'][];
 
@@ -173,7 +111,7 @@ export async function runKontrolProfileViaKaaS(
   }
 
   const body: components['schemas']['CreateJobDto'] = {
-    branch: gitInfo.branch,
+    branch: branch,
     kontrolVersion: '',
     kontrolDockerImage: '',
     kaasCliBranch: 'master',
